Move MDI mode switch to gcode/ocode helpers
Fixes #33
